# sample-service/src/main/java/t1/openschool/sampleservice/dto/UserPatchDto.java
package t1.openschool.sampleservice.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Dto for patching existing user")
public class UserPatchDto {
    private String name;
    private Integer age;
}