# sample-service/src/main/java/t1/openschool/sampleservice/dto/UserResponseDto.java
package t1.openschool.sampleservice.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Dto for response user info")
public class UserResponseDto {
    private Long id;
    private String name;
    private Integer age;
}